archive/old_implementation/requirements.txt
# Core Trading Dependencies
selenium>=4.15.0
python-dotenv>=1.0.0
eth-account>=0.10.0

# Hyperliquid SDK
hyperliquid-python-sdk>=0.2.0

# AI/ML Dependencies
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
matplotlib>=3.7.0  # For chart generation

# Optional: Advanced AI
anthropic>=0.18.0  # For Claude API integration
xgboost>=2.0.0  # For advanced ML models
plotly>=5.14.0  # For interactive HTML visualizations

# Optional: Deep Learning (uncomment if needed)
# torch>=2.1.0
# tensorflow>=2.15.0
